requirements.txt
praw
openai
torch
transformers
pandas
matplotlib
seaborn
plotly
python-dotenv
scikit-learn